# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NAUB Home Finder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/google:
    post:
      operationId: googleAuth
      tags: [auth]
      summary: Sign in or register with Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthInput"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/kyc/submit:
    post:
      operationId: submitKyc
      tags: [auth]
      summary: Submit KYC documents for landlord/agent verification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitInput"
      responses:
        "200":
          description: KYC submitted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── USERS ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterInput:
      type: object
      required:
        - email
        - password
        - role
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, landlord, agent, escrow_officer]
        first_name:
          type: ["string", "null"]
        last_name:
          type: ["string", "null"]
        matriculation_number:
          type: ["string", "null"]
    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required:
        - message
        - user
        - token
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── Users ────────────────────────────────────────────────────────────────
    User:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        first_name:
          type: ["string", "null"]
        last_name:
          type: ["string", "null"]
        matriculation_number:
          type: ["string", "null"]
        verification_status:
          type: ["string", "null"]
        account_suspended:
          type: ["boolean", "null"]
        phone_number:
          type: ["string", "null"]
        profile_photo_url:
          type: ["string", "null"]
        created_at:
          type: ["string", "null"]
...
    GoogleAuthInput:
      type: object
      required:
        - credential
      properties:
        credential:
          type: string
          description: Google ID token from the Google Sign-In button
        role:
          type: string
          enum: [student, landlord, agent]
          description: Desired role when creating a new account (defaults to student)
    KycSubmitInput:
      type: object
      required:
        - national_id_type
        - national_id_document_url
        - selfie_url
      properties:
        national_id_type:
          type: string
          enum: [nin, international_passport, drivers_licence]
        national_id_document_url:
          type: string
          description: Base64-encoded ID document image
        selfie_url:
          type: string
          description: Base64-encoded selfie image
        property_document_url:
          type: string
          description: Base64-encoded property ownership document (required for landlords)

    # ── Shared ────────────────────────────────────────────────────────────────
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string