# lib/api-zod/src/generated/types/googleAuthInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NAUB Home Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { GoogleAuthInputRole } from './googleAuthInputRole';

export interface GoogleAuthInput {
  /** Google ID token from the Google Sign-In button */
  credential: string;
  /** Desired role when creating a new account (defaults to student) */
  role?: GoogleAuthInputRole;
}