# lib/api-zod/src/generated/types/kycSubmitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NAUB Home Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { KycSubmitInputNationalIdType } from './kycSubmitInputNationalIdType';

export interface KycSubmitInput {
  national_id_type: KycSubmitInputNationalIdType;
  /** Base64-encoded ID document image */
  national_id_document_url: string;
  /** Base64-encoded selfie image */
  selfie_url: string;
  /** Base64-encoded property ownership document (required for landlords) */
  property_document_url?: string;
}